Implement Scoring & Respawn
Added Widget blueprints to show each players name, from the built in playerstate, and their score.
Score is handled via the TankGameState.
Each time a projectile hits another tank, it is destroyed, a score is added and a new tank is spawned.
We then use SetViewTargetWith blend on the client playercontroller and Possess.